more login updates
