+++ routes/cart.js
@@ -2,7 +2,7 @@ const express = require("express")
 const router = express.Router()
 
 router.get('/', function(req, res) {
-    res.render('cart', { title: 'Cart' });
+    res.render('cart', { title: 'Cart', loggedIn: req.session.loggedIn });
 });
 
 module.exports = router
