+++ routes/categories.js
@@ -2,7 +2,7 @@ const express = require("express")
 const router = express.Router()
 
 router.get('/', function(req, res) {
-    res.render('categories', { title: 'Etsy Clone' });
+    res.render('categories', { title: 'Etsy Clone', loggedIn: req.session.loggedIn });
 });
 
 module.exports = router
