+++ routes/product.js
@@ -2,7 +2,7 @@ const express = require("express")
 const router = express.Router()
 
 router.get('/', function(req, res) {
-    res.render('product', { title: 'Etsy Clone' });
+    res.render('product', { title: 'Etsy Clone', loggedIn: req.session.loggedIn });
 });
 
 module.exports = router
