+++ routes/store.js
@@ -2,7 +2,7 @@ const express = require("express")
 const router = express.Router()
 
 router.get('/', function(req, res) {
-    res.render('store', { title: 'Etsy Clone' });
+    res.render('store', { title: 'Etsy Clone', loggedIn: req.session.loggedIn });
 });
 
 module.exports = router
