+++ views/login.mustache
@@ -40,7 +40,7 @@
                                      data-client_id={{client_id}}
                                      data-context="signin"
                                      data-ux_mode="popup"
-                                     data-login_uri="http://localhost:3000/index"
+                                     data-login_uri="http://localhost:3000/loggingIn"
                                      data-auto_select="true">
                                 </div>
                                 <div class="g_id_signin"
